:memo: Add summary of public speaking advice video
This commit adds a new summary with some public speaking notes from
Julian Treasure on how to speak so that people want to listen to you.
These are the subtle ways in how we can be more effective as public
speakers by being more conscious of our sound and presentation.

Signed-off-by: Justin W. Flory <git@jwf.io>