:relaxed: Spring cleaning: Reorganize and recategorize
This commit takes care of some periodic hierarchy reorganization. I
purged the immigration notes, moved my thought bookmarks to a
"scratchpad" folder, along with Karl Fogel.

Fedora page is removed in this commit too because git was being stubborn
about adding that path back to the staging area, so it is going into
this commit too.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>